refactor(agent-core-v2): introduce strong TurnId
Unify runtime turn identifiers behind the TurnId newtype, preserve legacy wire decoding, and align persisted turn lifecycle operations with the TypeScript implementation.

+++ crates/agent-core-v2/src/agent/user_tool/service.rs
@@ -73,7 +73,7 @@ impl ExecutableTool for UserToolExecutable {
                     }),
                     origin: Some(InteractionOrigin {
                         agent_id: None,
-                        turn_id: Some(context.turn_id as f64),
+                        turn_id: Some(context.turn_id),
                     }),
                 };
                 match abortable(interaction.request(request), &context.signal).await {
